♻️ refactor(notification): 알림 설정 로직 리팩터링 및 모델 통합
- `AlarmSetting` 도메인 모델에 기본값 설정 및 모든 서브 알림 비활성 여부 확인 메서드 추가
- UI 상태 관리를 위해 사용하던 `AlarmToggleUiState`를 제거하고 `AlarmSetting` 도메인 모델로 통합
- `NotificationViewModel` 내 낙관적 업데이트 로직 및 전체 알림 토글 로직을 더 직관적으로 개선
- `AlarmRepository`의 알림 설정 업데이트 반환 타입을 `Result<Boolean>`으로 수정
- `NotificationPreference`에서 불필요한 개별 설정 키와 복잡한 동기화 로직을 제거하고 마스터 알림 설정만 유지하도록 간소화
